Address code review comments.
