+++ src/Razor/test/Microsoft.CodeAnalysis.Remote.Razor.Test/RemoteTagHelperDeltaProviderTest.cs
@@ -5,42 +5,11 @@
 
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.CodeAnalysis.Remote.Razor.Test;
 using Xunit;
 
-namespace Microsoft.CodeAnalysis.Remote.Razor.Test
+namespace Microsoft.CodeAnalysis.Remote.Razor
 {
-    public class TagHelperDescriptorTestBase
-    {
-        public TagHelperDescriptorTestBase()
-        {
-            Project1FilePath = "C:/path/to/Project1/Project1.csproj";
-            TagHelper1_Project1 = TagHelperDescriptorBuilder.Create("TagHelper1", "Project1").Build();
-            TagHelper2_Project1 = TagHelperDescriptorBuilder.Create("TagHelper2", "Project1").Build();
-
-            Project2FilePath = "C:/path/to/Project2/Project2.csproj";
-            TagHelper1_Project2 = TagHelperDescriptorBuilder.Create("TagHelper1", "Project2").Build();
-            TagHelper2_Project2 = TagHelperDescriptorBuilder.Create("TagHelper2", "Project2").Build();
-        }
-
-        protected string Project1FilePath { get; }
-
-        protected TagHelperDescriptor TagHelper1_Project1 { get; }
-
-        protected TagHelperDescriptor TagHelper2_Project1 { get; }
-
-        protected IReadOnlyList<TagHelperDescriptor> Project1TagHelpers => new[] { TagHelper1_Project1, TagHelper2_Project1 };
-
-        protected string Project2FilePath { get; }
-
-        protected TagHelperDescriptor TagHelper1_Project2 { get; }
-
-        protected TagHelperDescriptor TagHelper2_Project2 { get; }
-
-        protected IReadOnlyList<TagHelperDescriptor> Project2TagHelpers => new[] { TagHelper1_Project2, TagHelper2_Project2 };
-
-        protected IReadOnlyList<TagHelperDescriptor> Project1AndProject2TagHelpers => new[] { TagHelper1_Project1, TagHelper2_Project1, TagHelper1_Project2, TagHelper2_Project2 };
-    }
-
     public class RemoteTagHelperDeltaProviderTest : TagHelperDescriptorTestBase
     {
         public RemoteTagHelperDeltaProviderTest()
